Implement initial (poor) identifier/operator distinction, including libraries.
Identifiers are traditional alphanumeric, but allow trailing '?' or
'!' IFF they are right-delimited.

Operators are punctuation-only for some set of punctuation characters.

All of the special cases here are going away in favor of something a
good bit closer to Haskell, per the note that I am sending to BitC-Dev
momentarily.