Handle unary negation properly in block syntax.
In the process: differentiate natural and negative numbers at the
parse level, so that we can ensure that types only get non-negtive
values where they need sizes. Need to preserve negative integers for
now, but only until we drop the s-expression syntax (might still need
it then for repr types, but I don't think so).