Change layout rules, add support for autoblock on THEN, ELSE clauses.
Curly is now auto-inserted by lexer after THEN, ELSE. It is closed by
corresponding parse error recovery rules.

This change caused the following code to break:

        let want = { if (on) then s.bits[w] | bit
                             else s.bits[w] & ~bit } in

because the explicit closing } is taken to be closing the inserted {
after ELSE. My diagnosis here is that the Haskell layout rules work
well for vertical stuff, and not so well for horizontal stuff. I added
a rule that explicit } has the effect of closing all implicit blocks
back to the matching explicit {. The new rule fixes this case and
several others.