Make most { } optional and recoverable.
This implements the push-back strategy described in the design note. A
bit tricky, since I had to record lastTokenType in a way that
preserves the chain for the layout layer, but doable.

The one set of {} that isn't optional are the ones that differentiate
type definitions from type declarations, because there is an inherent
ambiguity in that case.