Update INOstream to support a beginning-of-line marker.
This update adds support in INOstream to have a common leading string
emitted after the indent and before the actual content of the
line. This allows C++-style block comments to be emitted, which became
useful when I added some debugging comment emission into the
pretty-printer to track down an issue there.

Updated gen-c.cxx such that it now emits the S-expr tree within a
C++-style block comment rather than a C-style comment. This was
needed so that C-style comments emitted within the pretty printer
(which don't nest) would not break the emitted C code.