Revise core exception structure to report source file: line of THROW.
This version is more useful for debugging purposes than a report
that an uncaught exception was raised from an unspecified location.:-)

There is a missing piece: intrinsic functions report the file and line
number within the intrinsic C code rather than the file and line
number in the user's source code from which the call was made.

I should probably update this to report the offset within line as
well, since we have it, and that's what every other error message in
the compiler reports.

The current implementation relies on modifying the exception structure
to carry additional information. This makes the exception structure
non-constant for the statically allocated exceptions, which is
SMP-unsafe. I need to move the exception location information into a
more suitable place, at which time some of this will get backed out.